refactor(utilities): rename `estados` to `states`
